+++ build-httplib/CMakeFiles/test_amalgamation.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/test_amalgamation.dir/tests/test_amalgamation.cpp.o"
+  "CMakeFiles/test_amalgamation.dir/tests/test_amalgamation.cpp.o.d"
+  "test_amalgamation"
+  "test_amalgamation.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/test_amalgamation.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()